% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/widgety.R
\name{renderWidgety}
\alias{renderWidgety}
\title{Widget render function for use in Shiny}
\usage{
renderWidgety(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

